Add expense detail components and refactor packages
Add new expense detail UI components and primitives (Hero, GeneralInfo, Currency, SplitBreakdown, AddOnsSummary, CashTranchesDetail, Notes, Provenance, and supporting primitives). Reorganize many existing components into form/ subpackages (chips, addon, payment, receipt, split) and update imports/usages (preview, steps, and screens) to match the new structure.
