+++ features/expenses/src/main/kotlin/es/pedrazamiguez/splittrip/features/expense/presentation/component/detail/CashTranchesDetailSection.kt
@@ -1,0 +1,57 @@
+package es.pedrazamiguez.splittrip.features.expense.presentation.component.detail
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import es.pedrazamiguez.splittrip.core.designsystem.foundation.spacing
+import es.pedrazamiguez.splittrip.core.designsystem.icon.TablerIcons
+import es.pedrazamiguez.splittrip.core.designsystem.icon.outline.Wallet
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.layout.FlatCard
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.text.BodyText
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.text.LabelText
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.text.SecondaryBodyText
+import es.pedrazamiguez.splittrip.features.expense.R
+import es.pedrazamiguez.splittrip.features.expense.presentation.model.CashTrancheDetailUiModel
+import kotlinx.collections.immutable.ImmutableList
+
+@Composable
+internal fun CashTranchesDetailSection(
+    tranches: ImmutableList<CashTrancheDetailUiModel>,
+    sourceCurrency: String
+) {
+    Column(verticalArrangement = Arrangement.spacedBy(MaterialTheme.spacing.Small)) {
+        Row(
+            horizontalArrangement = Arrangement.spacedBy(6.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            SectionIcon(TablerIcons.Outline.Wallet)
+            LabelText(text = stringResource(R.string.add_expense_cash_tranche_funded_from))
+        }
+        FlatCard(modifier = Modifier.fillMaxWidth()) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(MaterialTheme.spacing.Default),
+                verticalArrangement = Arrangement.spacedBy(MaterialTheme.spacing.Small)
+            ) {
+                tranches.forEach { tranche ->
+                    Row(
+                        modifier = Modifier.fillMaxWidth(),
+                        horizontalArrangement = Arrangement.SpaceBetween
+                    ) {
+                        SecondaryBodyText(text = sourceCurrency)
+                        BodyText(text = tranche.formattedAmountConsumed)
+                    }
+                }
+            }
+        }
+    }
+}
