+++ features/expenses/src/main/kotlin/es/pedrazamiguez/splittrip/features/expense/presentation/component/detail/GeneralInfoSection.kt
@@ -1,0 +1,56 @@
+package es.pedrazamiguez.splittrip.features.expense.presentation.component.detail
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import es.pedrazamiguez.splittrip.core.designsystem.foundation.spacing
+import es.pedrazamiguez.splittrip.core.designsystem.icon.TablerIcons
+import es.pedrazamiguez.splittrip.core.designsystem.icon.outline.CreditCard
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.layout.FlatCard
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.text.LabelText
+import es.pedrazamiguez.splittrip.features.expense.R
+import es.pedrazamiguez.splittrip.features.expense.presentation.model.ExpenseDetailUiModel
+
+@Composable
+internal fun GeneralInfoSection(expense: ExpenseDetailUiModel) {
+    Column(verticalArrangement = Arrangement.spacedBy(MaterialTheme.spacing.Small)) {
+        Row(
+            horizontalArrangement = Arrangement.spacedBy(6.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            SectionIcon(TablerIcons.Outline.CreditCard)
+            LabelText(text = stringResource(R.string.expense_detail_section_info))
+        }
+        FlatCard(modifier = Modifier.fillMaxWidth()) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(MaterialTheme.spacing.Default),
+                verticalArrangement = Arrangement.spacedBy(MaterialTheme.spacing.Medium)
+            ) {
+                DetailRow(
+                    label = stringResource(R.string.expense_review_payment_method),
+                    value = expense.paymentMethodText
+                )
+                DetailRow(
+                    label = stringResource(R.string.expense_detail_paid_by_label),
+                    value = expense.createdByText
+                )
+                if (expense.vendorText != null) {
+                    DetailRow(
+                        label = stringResource(R.string.expense_review_vendor),
+                        value = expense.vendorText
+                    )
+                }
+            }
+        }
+    }
+}
