+++ features/expenses/src/main/kotlin/es/pedrazamiguez/splittrip/features/expense/presentation/component/step/expense/PaymentStatusStep.kt
@@ -4,8 +4,8 @@ import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.wizard.WizardStepLayout
-import es.pedrazamiguez.splittrip.features.expense.presentation.component.DueDateSection
-import es.pedrazamiguez.splittrip.features.expense.presentation.component.PaymentStatusSection
+import es.pedrazamiguez.splittrip.features.expense.presentation.component.form.DueDateSection
+import es.pedrazamiguez.splittrip.features.expense.presentation.component.form.payment.PaymentStatusSection
 import es.pedrazamiguez.splittrip.features.expense.presentation.viewmodel.event.AddExpenseUiEvent
 import es.pedrazamiguez.splittrip.features.expense.presentation.viewmodel.state.AddExpenseUiState
