+++ features/expenses/src/main/kotlin/es/pedrazamiguez/splittrip/features/expense/presentation/component/step/expense/ReceiptStep.kt
@@ -3,7 +3,7 @@ package es.pedrazamiguez.splittrip.features.expense.presentation.component.step.
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.wizard.WizardStepLayout
-import es.pedrazamiguez.splittrip.features.expense.presentation.component.ReceiptSection
+import es.pedrazamiguez.splittrip.features.expense.presentation.component.form.receipt.ReceiptSection
 import es.pedrazamiguez.splittrip.features.expense.presentation.viewmodel.event.AddExpenseUiEvent
 import es.pedrazamiguez.splittrip.features.expense.presentation.viewmodel.state.AddExpenseUiState
